backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
PyJWT==2.12.1
redis==6.4.0
requests==2.32.3
sqlparse==0.5.5
tzdata==2025.3
waitress==3.0.0
